narrow free variables (again) whenever they are demanded.
instead of sharing its evaluated non-deterministic generator,
a free variable can be narrowed again w.r.t. the current constraint store
when it is demanded. Recomputing should not be much slower than traversing
the choices and getting failures on possibly many branches.

With this approach, the `Narrow` instances can generate only those values
for free variables that are still valid w.r.t. the current constraint store.
They can, hence, avoid generating unecessary choices.

This seems to be especially useful for custom constraint solvers on
specific domains (like integers) but can also be done with 'narrowed' variables.
Whether it is done, depends on the definition of `narrow` in the `Narrow`
instance for a specific type.