added narrowing policies.
variables can either be narrowed on creation and shared on every demand
or renarrowed whenever they are demanded according to the current
constraint store.

The default is narrowing on creation which is better for simple
non-deterministic generators that don't consider the current constraints.
Using the `OnDemand` policy, one can avoid unnessesary branchings and
yield only those results that are consistent with the current constrint
store.