fixed `narrow` definition.
When defining instances of `Narrow`, the `narrow` function *must*
supply the given unique identifier *as is* to the call to `oneOf`
rather than generating a fresh id for this call.

Reason: the `FreeVar` constructor stores the value of the supply
that is passed to `oneOf`. Using this supply is safe, because `oneOf`
does not split the given supply but only uses its identifier.